Move NewChatDropDownBox.jsx to components/
