+++ frontend/src/components/NewChatDropDownBox.jsx
@@ -1,6 +1,6 @@
 import { motion } from "framer-motion";
 
-import Button from "../ui/Button";
+import Button from "./ui/Button";
 
 const DropDownMenuItem = ({ menuItem }) => {
   return (
